Using endpoints in worker with new, dedicated, integration test
This is still work in progress as the few notes in docstrings
indicate.

Twisted's PB classes are still too much entangled with ClientFactory
to get completely rid of that. On the other hand, pbutil's
reconnecting factory is meant to add a layer of retry logic
(the hung timeout).

Remaining work:
- the hung timeout is not yet fully reimplemented
  (it wouldn't work after a reconnect) and not tested
- the maxDelay is ignored (should be forwarded to service
  instantiation)
- graceful shutdowns aren't tested here, but it could be that
  unit/test_bot_Worker.py, which could also qualify as
  integration test does it enough.
- docstrings must be cleaned up and use current syntax (is it
  Sphinx's?) for attributes, methods, etc.
- a bit more of unit testing for connection description string
  and host escaping.